code improvements and styling

+++ src/app/services/firestore.service.ts
@@ -6,7 +6,7 @@ import {
 } from '@angular/fire/compat/firestore';
 
 import { Observable } from 'rxjs';
-import { switchMap } from 'rxjs/operators';
+import { map, switchMap } from 'rxjs/operators';
 
 import { AuthenticationService } from 'src/app/services/authentication.service';
 
@@ -53,29 +53,70 @@ export class FirestoreService {
     );
   }
 
-  async getDoc(docID: any) {
+  async getDocc(docID: string) {
     const docRef = await this.movementsCollectionRef?.doc(docID);
     docRef?.get().subscribe((value) => console.log(value.data()));
   }
 
-  async updateDoc(docID: any) {
-    await this.movementsCollectionRef
-      ?.doc(docID)
-      .update({ prop: 'updatedStrinGonDOC' });
-    console.log(`Movement with id ${docID} is updated`);
+  updateDoc(collection: string, docID: string) {
+    if (collection === 'categories') {
+      this.categoriesCollectionRef?.doc(docID).update({});
+      console.log(`Category with id ${docID} is updated`);
+    }
+    if (collection === 'movements') {
+      this.movementsCollectionRef?.doc(docID).update({});
+      console.log(`Movement with id ${docID} is updated`);
+    }
   }
 
-  async deleteDoc(docID: any) {
-    await this.movementsCollectionRef?.doc(docID).delete();
-    console.log(`Movement with id ${docID} is deleted`);
+  deleteDoc(collection: string, docID: string) {
+    if (collection === 'categories') {
+      this.categoriesCollectionRef?.doc(docID).delete();
+      console.log(`Category with id ${docID} is deleted`);
+    }
+    if (collection === 'movements') {
+      this.movementsCollectionRef?.doc(docID).delete();
+      console.log(`Movement with id ${docID} is deleted`);
+    }
   }
 
-  updateCategory(docID: string) {
-    this.categoriesCollectionRef?.doc(docID).update({});
-    console.log(`Category with id ${docID} is updated`);
-  }
-  deleteCategory(docID: string) {
-    this.categoriesCollectionRef?.doc(docID).delete();
-    console.log(`Category with id ${docID} is deleted`);
+  filterAndSortDocs<T>(
+    collection: Observable<T[]>,
+    filterAndSortBy: {
+      type: string;
+      sortBy: string;
+    }
+  ): Observable<T[]> {
+    if (filterAndSortBy.type === 'all') {
+      return collection.pipe(
+        map((docs: T[]) =>
+          docs.sort((a: any, b: any) =>
+            filterAndSortBy.sortBy === 'date'
+              ? b[filterAndSortBy.sortBy].localeCompare(
+                  a[filterAndSortBy.sortBy]
+                )
+              : a[filterAndSortBy.sortBy].localeCompare(
+                  b[filterAndSortBy.sortBy]
+                )
+          )
+        )
+      );
+    } else {
+      return collection.pipe(
+        map((docs: T[]) =>
+          docs
+            .sort((a: any, b: any) =>
+              filterAndSortBy.sortBy === 'date'
+                ? b[filterAndSortBy.sortBy].localeCompare(
+                    a[filterAndSortBy.sortBy]
+                  )
+                : a[filterAndSortBy.sortBy].localeCompare(
+                    b[filterAndSortBy.sortBy]
+                  )
+            )
+            .filter((doc: any) => doc.type === filterAndSortBy.type)
+        )
+      );
+    }
   }
 }
